app/server/pyproject.toml
[tool.poetry]
name = "fydp-2025-g55"
version = "0.1.0"
description = ""
authors = ["Vivek Alamuri <valamuri@uwaterloo.ca>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
isort = "^5.13.2"
boto3 = "^1.34.144"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
